misc: setup and start fixing Clippy lints

+++ runtime/src/classes/java/lang/boxes.rs
@@ -31,10 +31,10 @@ macro_rules! primitive_boxes {
 primitive_boxes!(
 	Boolean,   java_lang_Boolean,   jboolean, FieldType::Boolean,   expect_int,    field, field != 0;
 	Character, java_lang_Character, jchar,    FieldType::Character, expect_int,    field, field as jchar;
-	Float,     java_lang_Float,     jfloat,   FieldType::Float,     expect_float,  field, field as jfloat;
-	Double,    java_lang_Double,    jdouble,  FieldType::Double,    expect_double, field, field as jdouble;
+	Float,     java_lang_Float,     jfloat,   FieldType::Float,     expect_float,  field, field;
+	Double,    java_lang_Double,    jdouble,  FieldType::Double,    expect_double, field, field;
 	Byte,      java_lang_Byte,      jbyte,    FieldType::Byte,      expect_int,    field, field as jbyte;
 	Short,     java_lang_Short,     jshort,   FieldType::Short,     expect_int,    field, field as jshort;
-	Integer,   java_lang_Integer,   jint,     FieldType::Integer,   expect_int,    field, field as jint;
-	Long,      java_lang_Long,      jlong,    FieldType::Long,      expect_long,   field, field as jlong;
+	Integer,   java_lang_Integer,   jint,     FieldType::Integer,   expect_int,    field, field;
+	Long,      java_lang_Long,      jlong,    FieldType::Long,      expect_long,   field, field;
 );
